memoize TxRow
